+++ packages/andrjs/src/api/ADODatabaseAPI.ts
@@ -64,6 +64,19 @@ export default class ADODatabaseAPI extends ADOAPI {
   }
 
   /**
+   * Provides a message object for the ADO DB's `GetCodeId` query
+   * @param name
+   * @returns
+   */
+  getAdoTypeQuery(codeId: number) {
+    return {
+      ado_type: {
+        code_id: codeId,
+      },
+    };
+  }
+
+  /**
    * Gets the code ID for an ADO type from the ADO DB
    * @param name
    * @param address
@@ -79,13 +92,31 @@ export default class ADODatabaseAPI extends ADOAPI {
   }
 
   /**
+ * Gets the code ID for an ADO type from the ADO DB
+ * @param name
+ * @param address
+ * @returns
+ */
+  async getAdoType(codeId: number, address?: string) {
+    if (!this.address && !address)
+      throw new Error("No provided ADO DB address to retrieve code ID");
+
+    const msg = this.getAdoTypeQuery(codeId);
+
+    return this.client.queryContract<string>(address ?? this.address, msg);
+  }
+
+  /**
    * Provides a message object for the ADO DB's `GetCodeId` query
    * @param name
    * @returns
    */
-  getAllADOQuery() {
+  getAllADOQuery(startAfter = '', limit = 100) {
     return {
-      all_ado_types: {},
+      all_ado_types: {
+        limit,
+        start_after: startAfter
+      },
     };
   }
 
@@ -95,11 +126,11 @@ export default class ADODatabaseAPI extends ADOAPI {
    * @param address
    * @returns
    */
-  async getAllADO(address?: string) {
+  async getAllADO(startAfter = '', limit = 100, address?: string) {
     if (!this.address && !address)
       throw new Error("No provided ADO DB address to retrieve code ID");
 
-    const msg = this.getAllADOQuery();
+    const msg = this.getAllADOQuery(startAfter, limit);
 
     return this.client.queryContract<string[]>(address ?? this.address, msg);
   }