+++ packages/andrjs/src/api/ADOSchemaAPI.ts
@@ -1,24 +1,14 @@
 import { Schema } from "jsonschema";
 import ADOAPI from "./ADOAPI";
 import AndromedaClient, { ContractSchema } from "index";
+import axios from "axios";
 
-export default class ADODSchemaAPI extends ADOAPI {
-  constructor(client: AndromedaClient, public address: string) {
+export default class ADOSchemaAPI extends ADOAPI {
+  constructor(client: AndromedaClient, public address: string = "") {
     super(client, address);
   }
 
-  /**
-   * Provides a message object for the Primitive DB's `GetKeyValue` query
-   * @param key
-   * @returns
-   */
-  getValueQuery(key: string) {
-    return {
-      get_value: {
-        key: key,
-      },
-    };
-  }
+  public static readonly SCHEMA_BASE_URL = 'https://api.andromedaprotocol.io/v1/schema'
 
   /**
    * Gets the schema for the key in schemadb
@@ -26,17 +16,13 @@ export default class ADODSchemaAPI extends ADOAPI {
    * @param address
    * @returns
    */
-  async getSchemaFromKey<T = ContractSchema>(key: string, address?: string) {
-    const msg = this.getValueQuery(key);
-    const res = await this.client.queryContract<{
-      key: string;
-      value: {
-        string: string;
-      }
-    }>(address ?? this.address, msg);
-    const schema = JSON.parse(res.value.string) as T;
+  private async getSchemaFromVersion<T = ContractSchema>(adoVersion: string, subSchema = 'default') {
+    const [adoType, version] = adoVersion.split('@');
+    const subPath = subSchema === 'default' ? adoType : subSchema;
+    const schema: T = await axios.get(`${ADOSchemaAPI.SCHEMA_BASE_URL}/raw/${adoType}/${version}/${subPath}`).then(res => res.data);
     return {
-      schema, key
+      schema,
+      adoVersion: adoVersion
     }
   }
 
@@ -47,40 +33,31 @@ export default class ADODSchemaAPI extends ADOAPI {
    * @param address
    * @returns
    */
-  async getSchemaFromCodeId(codeId: number, address?: string) {
-    return this.getSchemaFromKey(codeId.toString(), address);
+  async getSchemaFromCodeId<T = ContractSchema>(codeId: number, subSchema = 'default', fallbackType?: string) {
+    let schema = await axios.get(`${ADOSchemaAPI.SCHEMA_BASE_URL}/raw/code_id/${codeId}/${await this.client.chainClient?.queryClient?.getChainId()}/${subSchema}`).then(res => res.data as T).catch(() => undefined);
+    if (!schema) {
+      schema = await this.client.os.adoDB?.getAdoType(codeId).then(adoVersion => this.getSchemaFromVersion<T>(adoVersion, subSchema)).then(data => data.schema).catch(() => undefined);
+    }
+    // If we still don't have schema, try to get ado type from ado type query
+    if (!schema && fallbackType) {
+      schema = await this.getSchemaFromVersion<T>(fallbackType, subSchema).then(data => data.schema).catch(() => undefined);
+    }
+    if (!schema) throw new Error("Schema not found!");
+    return {
+      schema: schema!,
+      codeId,
+    }
   }
 
-  /**
- * Gets the schema for the adoType. First fetch the codeId from adodb and then use the codeId to get the schema
- * @param key
- * @param address
- * @returns
- */
-  async getSchemaFromAdoType(adoType: string, address?: string) {
-    const codeId = await this.client.os.adoDB!.getCodeId(adoType);
-    return this.getSchemaFromCodeId(codeId, address);
-  }
 
   /**
-  * Gets the sub schema, like receive schema, from the codeId. Similar to getSchemaFromCodeId
-  * @param key
-  * @param address
-  * @returns
-  */
-  async getSubSchemaFromCodeId(codeId: number, subKey: string, address?: string) {
-    const key = `${codeId}-${subKey}`
-    return this.getSchemaFromKey<Schema>(key, address);
+* Gets the sub schema, like receive schema, from the codeId. Similar to getSchemaFromCodeId
+* @param key
+* @param address
+* @returns
+*/
+  async getSubSchemaFromCodeId(codeId: number, subKey: string) {
+    return this.getSchemaFromCodeId<Schema>(codeId, subKey);
   }
 
-  /**
-  * Get the sub schema, like receive schema, from adoType. Similar to getSchemaFromAdoType
-  * @param key
-  * @param address
-  * @returns
-  */
-  async getSubSchemaFromAdoType(adoType: string, subKey: string, address?: string) {
-    const codeId = await this.client.os.adoDB!.getCodeId(adoType);
-    return this.getSubSchemaFromCodeId(codeId, subKey, address);
-  }
 }