+++ packages/andrjs/src/api/OperatingSystemAPI.ts
@@ -2,14 +2,11 @@ import { validateAddress } from "..";
 import type AndromedaClient from "../AndromedaClient";
 import ADOAPI from "./ADOAPI";
 import ADODatabaseAPI from "./ADODatabaseAPI";
-import ADODSchemaAPI from "./ADOSchemaAPI";
 import VirtualFileSystemAPI from "./VirtualFileSystemAPI";
 
 export default class OperatingSystemAPI extends ADOAPI {
   public vfs?: VirtualFileSystemAPI;
   public adoDB?: ADODatabaseAPI;
-  // API for shared SCHEMA messages
-  public schema?: ADODSchemaAPI;
 
   constructor(client: AndromedaClient, public address: string = "") {
     super(client, address);
@@ -35,7 +32,6 @@ export default class OperatingSystemAPI extends ADOAPI {
     this.address = address;
     await this.fetchVFSAddress();
     await this.fetchADODBAddress();
-    await this.fetchSchemaAddress();
   }
 
   /**
@@ -55,15 +51,6 @@ export default class OperatingSystemAPI extends ADOAPI {
     const adoDBKey = "adodb";
     const adoDBAddress = await this.fetchKeyAddress(adoDBKey);
     this.adoDB = new ADODatabaseAPI(this.client, adoDBAddress);
-  }
-
-  /**
- * Fetches the current ADOP Schema address from the kernel
- */
-  async fetchSchemaAddress() {
-    const key = "schemadb";
-    const address = await this.fetchKeyAddress(key);
-    this.schema = new ADODSchemaAPI(this.client, address);
   }
 
   /**