+++ packages/andrjs/src/clients/InjectiveClient.ts
@@ -1,126 +1,49 @@
 import {
   CosmWasmClient,
-  ExecuteResult,
   InstantiateOptions,
   InstantiateResult,
   MigrateResult,
+  SigningCosmWasmClientOptions,
+  UploadResult,
+  wasmTypes,
 } from "@cosmjs/cosmwasm-stargate";
-import { sha256 } from "@cosmjs/crypto";
-import { toHex } from "@cosmjs/encoding";
+
+import { Fee, Msg } from '..'
 import {
   Coin,
   EncodeObject,
-  OfflineDirectSigner as CosmosOfflineDirectSigner,
+  OfflineDirectSigner,
+  Registry,
 } from "@cosmjs/proto-signing";
-import { DeliverTxResponse, GasPrice, StdFee } from "@cosmjs/stargate";
-import { findAttribute, Log, parseRawLog } from "@cosmjs/stargate/build/logs";
+import { AminoTypes, DeliverTxResponse, GasPrice, StdFee, defaultRegistryTypes } from "@cosmjs/stargate";
+
 import {
-  getNetworkChainInfo,
-  getNetworkEndpoints,
-  Network,
-} from "@injectivelabs/networks";
-import {
-  BaseAccount,
-  ChainRestAuthApi,
-  ChainRestTendermintApi,
-  createTransaction,
-  MsgArg,
-  MsgExecuteContract,
-  MsgInstantiateContract,
-  MsgMigrateContract,
-  MsgSend,
-  MsgStoreCode,
-  TxGrpcClient,
-  TxRaw as InjTxRaw,
-  createTxRawFromSigResponse,
+  InjectiveStargate,
 } from "@injectivelabs/sdk-ts";
-import { OfflineDirectSigner } from "@injectivelabs/sdk-ts/dist/core/accounts/signers/types/proto-signer";
 import {
-  BigNumberInBase,
-  DEFAULT_BLOCK_TIMEOUT_HEIGHT,
-  DEFAULT_GAS_LIMIT,
-  getStdFee,
+  DEFAULT_FEE_DENOM,
+  DEFAULT_GAS_PRICE,
+  DEFAULT_STD_FEE,
 } from "@injectivelabs/utils";
-import _ from "lodash";
-import Long from "long";
-import { gzip } from "pako";
-import type { Msg } from "../types";
 import BaseChainClient from "./BaseChainClient";
 import ChainClient from "./ChainClient";
+import { SigningStargateClient, createDefaultAminoConverters } from "@injectivelabs/sdk-ts/dist/cjs/core/stargate/SigningStargateClient";
+import { TxRaw } from "cosmjs-types/cosmos/tx/v1beta1/tx";
+import { gzip } from "pako";
+import { sha256 } from "@cosmjs/crypto";
+import { toHex } from "@cosmjs/encoding";
+import { findAttribute, Log } from "@cosmjs/stargate/build/logs";
+import { Tendermint37Client } from "@cosmjs/tendermint-rpc";
 
-type MsgType =
-  | "MsgSend"
-  | "MsgExecuteContract"
-  | "MsgInstantiateContract"
-  | "MsgStoreCode"
-  | "MsgMigrateContract";
-
-function arrayToJson(array: Uint8Array) {
-  const jsonString = Buffer.from(array).toString("utf8");
-
-  const parsedData = JSON.parse(jsonString);
-  return parsedData;
-}
-
-function mapObjToEnjClass(type: MsgType, value: any) {
-  switch (type) {
-    case "MsgInstantiateContract":
-      return MsgInstantiateContract.fromJSON({
-        sender: value.sender,
-        admin: value.sender,
-        codeId: value.codeId,
-        label: value.label,
-        // Msg is encoded as Uint8Array, convert back to JSON here
-        msg: arrayToJson(value.msg),
-        amount: value.amount,
-      });
-    case "MsgExecuteContract":
-      return MsgExecuteContract.fromJSON({
-        funds: value.funds,
-        sender: value.sender,
-        contractAddress: value.contract,
-        msg: arrayToJson(value.msg),
-      });
-    case "MsgMigrateContract":
-      return MsgMigrateContract.fromJSON({
-        sender: value.sender,
-        contract: value.contract,
-        codeId: value.codeId,
-        msg: arrayToJson(value.msg),
-      });
-    case "MsgSend":
-      return MsgSend.fromJSON({
-        amount: value.amount[0],
-        srcInjectiveAddress: value.fromAddress,
-        dstInjectiveAddress: value.toAddress,
-      });
-    case "MsgStoreCode":
-    default:
-      return MsgStoreCode.fromJSON({
-        sender: value.sender,
-        wasmBytes: value.wasmByteCode,
-      });
-  }
-}
-
-function encodeObjectToMsgArgs(msgs: EncodeObject[]): MsgArg[] {
-  return msgs.map((msg) => {
-    const type = _.last(msg.typeUrl.split("."));
-    return mapObjToEnjClass(type as MsgType, msg.value).toDirectSign();
-  });
-}
 
 export default class InjectiveClient
   extends BaseChainClient
   implements ChainClient {
-  public signingClient?: TxGrpcClient;
-  public queryClient?: CosmWasmClient;
+  declare public signingClient?: SigningStargateClient;
+  declare public queryClient?: CosmWasmClient;
+  public gasPrice?: GasPrice;
 
-  private directSigner?: OfflineDirectSigner | CosmosOfflineDirectSigner;
-  private chainRestTendermintApi?: ChainRestTendermintApi;
-  private chainRestAuthApi?: ChainRestAuthApi;
-  private gasPrice?: GasPrice;
-  private chainId?: string;
+  private directSigner?: OfflineDirectSigner;
 
   protected preMessage(signed = true) {
     super.preMessage(signed);
@@ -131,22 +54,22 @@ export default class InjectiveClient
 
   async connect(
     endpoint: string,
-    signer?: CosmosOfflineDirectSigner | OfflineDirectSigner
+    signer?: OfflineDirectSigner,
+    options?: SigningCosmWasmClientOptions
   ): Promise<void> {
     delete this.signingClient;
     delete this.queryClient;
-    const network = endpoint.includes("testnet")
-      ? Network.TestnetK8s
-      : Network.MainnetK8s;
-    this.chainId = getNetworkChainInfo(network).chainId;
-    const { rest, rpc, grpc } = getNetworkEndpoints(network);
 
-    this.queryClient = await CosmWasmClient.connect(rpc!);
-    this.chainRestTendermintApi = new ChainRestTendermintApi(rest);
-    this.chainRestAuthApi = new ChainRestAuthApi(rest);
 
+    this.queryClient = await CosmWasmClient.connect(endpoint);
+    this.gasPrice = options?.gasPrice || GasPrice.fromString(DEFAULT_GAS_PRICE + DEFAULT_FEE_DENOM);
     if (signer) {
-      this.signingClient = new TxGrpcClient(grpc);
+      const tmClient = await Tendermint37Client.connect(endpoint);
+      this.signingClient = await InjectiveStargate.InjectiveSigningStargateClient.createWithSigner(tmClient, signer, {
+        registry: options?.registry || new Registry([...defaultRegistryTypes, ...wasmTypes]),
+        aminoTypes: options?.aminoTypes || new AminoTypes({ ...createDefaultAminoConverters() }),
+        gasPrice: this.gasPrice
+      });
       this.directSigner = signer;
 
       const [account] = await signer.getAccounts();
@@ -161,223 +84,134 @@ export default class InjectiveClient
     this.signer = "";
     delete this.directSigner;
     delete this.gasPrice;
-    delete this.chainRestTendermintApi;
-    delete this.chainRestAuthApi;
   }
 
-  private async getTimeoutHeight() {
-    if (!this.chainRestTendermintApi)
-      throw new Error("Chain Tendermint Rest API not connected");
-    const latestBlock = await this.chainRestTendermintApi.fetchLatestBlock();
-    const latestHeight = latestBlock.header.height;
-    return new BigNumberInBase(latestHeight)
-      .plus(DEFAULT_BLOCK_TIMEOUT_HEIGHT)
-      .toNumber();
-  }
 
-  private async getBaseAccount() {
-    if (!this.chainRestAuthApi)
-      throw new Error("Chain Auth Rest API not connected");
-    const accountDetailsResponse = await this.chainRestAuthApi.fetchAccount(
-      this.signer
-    );
-    const baseAccount = BaseAccount.fromRestApi(accountDetailsResponse);
-    return baseAccount.toAccountDetails();
-  }
-
-  private async getPubKey() {
-    const [account] = await this.directSigner!.getAccounts();
-
-    return Buffer.from(account.pubkey).toString("base64");
-  }
-
-  private async signInj(
-    messages: EncodeObject[],
-    fee: StdFee = getStdFee((DEFAULT_GAS_LIMIT * 2).toString()),
-    memo: string = "",
-    simulation = false
-  ) {
-    this.preMessage();
-    const timeoutHeight = await this.getTimeoutHeight();
-    const baseAccount = await this.getBaseAccount();
-    const pubKey = await this.getPubKey();
-    const { signDoc, txRaw } = createTransaction({
-      pubKey,
-      chainId: this.chainId!,
-      message: encodeObjectToMsgArgs(messages),
-      timeoutHeight,
-      sequence: baseAccount.sequence,
-      accountNumber: baseAccount.accountNumber,
-      memo,
-      fee,
-    });
-
-    if (simulation) return txRaw;
-
-    const signed = await this.directSigner!.signDirect(this.signer!, {
-      ...signDoc,
-      chainId: signDoc.getChainId(),
-      bodyBytes: signDoc.getBodyBytes_asU8(),
-      authInfoBytes: signDoc.getAuthInfoBytes_asU8(),
-      accountNumber: Long.fromInt(baseAccount.accountNumber),
-    });
-
-    return createTxRawFromSigResponse(signed);
-  }
-
-  async sign(messages: EncodeObject[], fee?: StdFee, memo?: string) {
-    const injTxRaw = await this.signInj(messages, fee, memo);
-    return {
-      bodyBytes: injTxRaw.getBodyBytes_asU8(),
-      authInfoBytes: injTxRaw.getAuthInfoBytes_asU8(),
-      signatures: injTxRaw.getSignaturesList_asU8(),
-    };
+  async sign(messages: EncodeObject[], fee?: StdFee, memo?: string): Promise<TxRaw> {
+    this.preMessage(true);
+    return await this.signingClient!.sign(this.signer, messages, fee || DEFAULT_STD_FEE, memo || "");
   }
 
   async broadcast(
-    tx: InjTxRaw,
-    timeoutMs = 60000,
+    tx: TxRaw,
   ): Promise<DeliverTxResponse> {
-    const resp = await this.signingClient!.broadcastBlock(tx, 3);
-    const result = await this.signingClient!.fetchTxPoll(resp.txHash, timeoutMs);
-    return {
-      // BUG: Fix types here
-      ...result as any,
-      transactionHash: result.txHash,
-    }
+    this.preMessage(true);
+    const txBytes = TxRaw.encode(tx).finish();
+    return await this.signingClient!.broadcastTx(txBytes)
   }
 
   async signAndBroadcast(
     messages: EncodeObject[],
     fee?: StdFee | "auto",
     memo?: string
-  ): Promise<DeliverTxResponse & { logs: readonly Log[] }> {
-    const signed = await this.signInj(
+  ): Promise<DeliverTxResponse & { logs: Log[] }> {
+    this.preMessage(true);
+    const res = await this.signingClient!.signAndBroadcast(
+      this.signer,
       messages,
-      fee === "auto" ? undefined : fee,
+      fee || DEFAULT_STD_FEE,
       memo
     );
-    const resp = await this.broadcast(signed);
     return {
-      ...resp,
-      logs: parseRawLog(resp.rawLog),
-    };
+      ...res,
+      logs: JSON.parse(res.rawLog ?? '[]')
+    }
+
+  }
+  async simulateMulti(
+    messages: EncodeObject[],
+    _fee: Fee = "auto",
+    memo?: string | undefined
+  ): Promise<number> {
+    this.preMessage();
+    return this.signingClient!.simulate(this.signer, messages, memo);
   }
 
-  // Saving this here for potential refactor for extra functionality
-  // async simulateMulti(messages: EncodeObject[], fee: StdFee, memo = "") {
-  //   const txRaw = await this.signInj(messages, fee, memo);
-  //   const resp = await this.signingClient!.simulate(txRaw);
-  //   // TODO: CHECK THIS RESPONSE OBJECT AS IT MAY BE INVALID
-  //   return {
-  //     ...resp,
-  //     result: {
-  //       ...resp.result,
-  //       data: Buffer.from(resp.result.data),
-  //       events: resp.result.eventsList.map((ev) => ({
-  //         ...ev,
-  //         attributes: ev.attributes.map((attr) => ({
-  //           key: Buffer.from(attr.key),
-  //           value: Buffer.from(attr.value),
-  //           index: false,
-  //         })),
-  //       })),
-  //     },
-  //     gasInfo: {
-  //       gasUsed: Long.fromNumber(resp.gasInfo.gasUsed),
-  //       gasWanted: Long.fromNumber(resp.gasInfo.gasWanted),
-  //     },
-  //   };
-  // }
-
-  async simulateMulti(messages: EncodeObject[], fee?: StdFee, memo = "") {
-    const txRaw = await this.signInj(messages, fee, memo, true);
-    const resp = await this.signingClient!.simulate(txRaw);
-    // TODO: CHECK THIS RESPONSE OBJECT AS IT MAY BE INVALID
-    return resp.gasInfo.gasUsed;
-  }
-
-  async simulate(message: EncodeObject, fee?: StdFee, memo = "") {
-    return this.simulateMulti([message], fee, memo);
+  async simulate(
+    message: EncodeObject,
+    _fee?: Fee | undefined,
+    memo?: string | undefined
+  ): Promise<number> {
+    this.preMessage();
+    return this.signingClient!.simulate(this.signer, [message], memo);
   }
 
   async execute(
     contractAddress: string,
     msg: Msg,
-    fee?: StdFee,
+    fee?: StdFee | undefined,
     memo?: string | undefined,
-    funds: Coin[] = []
-  ): Promise<ExecuteResult> {
-    const message = this.encodeExecuteMsg(contractAddress, msg, funds);
-    return await this.signAndBroadcast([message], fee, memo);
+    funds?: readonly Coin[] | undefined
+  ): Promise<DeliverTxResponse & { logs: any }> {
+    this.preMessage(true);
+    const encoded = this.encodeExecuteMsg(contractAddress, msg, [...funds || []]);
+    const res = await this.signAndBroadcast([encoded], fee, memo);
+    return {
+      ...res,
+      logs: []
+    }
   }
 
   async simulateExecute(
     address: string,
     msg: Msg,
     funds: Coin[],
-    fee?: StdFee,
+    _fee?: Fee,
     memo?: string | undefined
   ) {
     const message = this.encodeExecuteMsg(address, msg, funds);
-    return this.simulate(message, fee, memo);
+    return this.simulate(message, undefined, memo);
   }
 
-  async upload(
-    code: Uint8Array,
-    fee?: StdFee,
-    memo = ""
-  ): ReturnType<ChainClient["upload"]> {
+  async upload(code: Uint8Array, fee: StdFee | "auto" = "auto", memo?: string | undefined): Promise<UploadResult> {
+    this.preMessage();
     const compressed = gzip(code, { level: 9 });
-    const message = this.encodeUploadMessage(compressed);
-    const resp = await this.signAndBroadcast([message], fee, memo);
-
-    const codeIdAttr = findAttribute(resp.logs, "store_code", "code_id");
-
-    // const originalChecksum = toHex(sha256(code));
+    const encoded = this.encodeUploadMessage(compressed);
+    const res = await this.signAndBroadcast([encoded], fee, memo);
     const compressedChecksum = toHex(sha256(compressed));
+    const codeIdAttr = findAttribute([{ msg_index: 0, 'log': '', events: res.events }], "cosmwasm.wasm.v1.EventCodeStored", "code_id");
+    let codeId = -1;
+    try {
+      codeId = parseInt(JSON.parse(codeIdAttr.value));
+    } catch (err) { }
     return {
-      ...resp,
-      codeId: parseInt(codeIdAttr.value, 10),
+      ...res,
       originalSize: code.length,
       compressedSize: compressed.length,
-      checksum: compressedChecksum
-    };
+      checksum: compressedChecksum,
+      codeId: codeId,
+      logs: []
+    }
   }
 
   async simulateUpload(
     code: Uint8Array,
-    fee?: StdFee,
+    _fee?: Fee | undefined,
     memo?: string | undefined
   ): Promise<number | undefined> {
-    const compressed = gzip(code, { level: 9 });
-    const message = this.encodeUploadMessage(compressed);
-    return this.simulate(message, fee, memo);
+    const message = this.encodeUploadMessage(code);
+    return this.simulate(message, undefined, memo);
   }
 
   async instantiate(
     codeId: number,
     msg: Msg,
     label: string,
-    fee?: StdFee,
+    fee: StdFee | "auto" = "auto",
     options?: InstantiateOptions
   ): Promise<InstantiateResult> {
-    const message = this.encodeInstantiateMsg(codeId, msg, label);
-    const resp = await this.signAndBroadcast(
-      [message],
-      fee,
-      options ? options.memo : ""
-    );
+    this.preMessage(true);
+    const encoded = this.encodeInstantiateMsg(codeId, msg, label, options);
+    const res = await this.signAndBroadcast([encoded], fee, "Instantiate");
     const contractAddressAttr = findAttribute(
-      resp.logs,
+      res.logs,
       "wasm",
       "_contract_address"
     );
-
     return {
-      ...resp,
+      ...res,
       contractAddress: contractAddressAttr.value,
-    };
+    }
   }
 
   async simulateInstantiate(
@@ -395,19 +229,19 @@ export default class InjectiveClient
     contractAddress: string,
     codeId: number,
     msg: Msg,
-    fee?: StdFee,
-    memo?: string
+    fee: StdFee | "auto" = "auto",
+    memo?: string | undefined
   ): Promise<MigrateResult> {
-    const message = this.encodeMigrateMessage(contractAddress, codeId, msg);
-    const resp = await this.signAndBroadcast([message], fee, memo);
-    return resp;
+    this.preMessage(true);
+    const encoded = this.encodeMigrateMessage(contractAddress, codeId, msg);
+    return this.signAndBroadcast([encoded], fee, memo);
   }
 
   async simulateMigrate(
     contractAddress: string,
     codeId: number,
     msg: Msg,
-    fee?: StdFee,
+    fee?: Fee | undefined,
     memo?: string | undefined
   ): Promise<number | undefined> {
     const message = this.encodeMigrateMessage(contractAddress, codeId, msg);
@@ -417,16 +251,15 @@ export default class InjectiveClient
   async sendTokens(
     receivingAddress: string,
     amount: readonly Coin[],
-    fee?: StdFee,
-    memo?: string
-  ): Promise<DeliverTxResponse> {
-    if (amount.length > 1)
-      throw new Error(
-        "Injective only enables the sending of one amount at a time, please only send one token type."
-      );
-
-    const message = this.encodeSendMessage(receivingAddress, [...amount]);
-
-    return this.signAndBroadcast([message], fee, memo);
+    fee: Fee = "auto",
+    memo?: string | undefined
+  ): Promise<any> {
+    return this.signingClient?.sendTokens(
+      this.signer,
+      receivingAddress,
+      amount,
+      fee,
+      memo
+    );
   }
 }