+++ packages/cli/src/handlers/ado/index.ts
@@ -44,20 +44,6 @@ const commands: Commands = {
     inputs: [
       {
         requestMessage: "Input the ADO type:",
-        validate: async (input: string) => {
-          try {
-            await client.os.schema!.getSchemaFromAdoType(input);
-            return true;
-          } catch (error) {
-            const { message } = error as Error;
-            if (message.includes("u64 not found")) {
-              console.log(pc.red("Invalid ADO Type"));
-            } else {
-              console.log(pc.red(message));
-            }
-            return false;
-          }
-        },
         options: async () => {
           try {
             const adoTypes = displaySpinnerAsync(
@@ -242,15 +228,14 @@ async function createHandler(input: string[], flags: Flags) {
   const [type] = input;
   const codeId = await client!.os!.adoDB!.getCodeId(type);
   const adoSchema = await displaySpinnerAsync(
-    `Fetching schema for ${type}-${codeId}...`,
-    async () => await client.os.schema!.getSchemaFromCodeId(codeId)
+    `Fetching schema for ${type} (${codeId}) ...`,
+    async () => await client.schema!.getSchemaFromCodeId(codeId)
   );
 
   const msg = await promptInstantiateMsg(
     adoSchema.schema.instantiate
       ? adoSchema.schema.instantiate
       : (adoSchema.schema as Schema),
-    codeId
   );
   await instantiateMessage(codeId, msg, flags);
 }
@@ -266,40 +251,23 @@ async function queryCodeId(address: string) {
 }
 
 /**
- * Queries an ADO for its type by address
- * @param address The address of the ADO
- * @returns The type of ADO the contract is, errors if the contract is not an ADO
- */
-async function queryADOType(address: string) {
-  const queryMsg = client.ado.typeQuery();
-
-  const { ado_type } = await queryMessage<{ ado_type: string }>(
-    address,
-    queryMsg
-  );
-
-  return ado_type;
-}
-
-/**
  * Queries an ADO for its schema by address
  * @param address The address of the ADO
  * @returns Schema for the ado
  */
 async function queryAdoSchema(address: string) {
-  try {
-    // First try to get the schema using codeId as codeId are unique
-    const codeId = await queryCodeId(address);
-    const schema = await client.os.schema!.getSchemaFromCodeId(codeId);
-    return schema;
-  } catch (err) {
-    // If codeId schema fetch fail, try to get the adoType and check the codeId for it in adodb
-    console.log(`Schema not found using Code Id, falling back to Ado Type`)
-    const adoType = await queryADOType(address);
-    console.log(adoType);
-    const schema = await client.os.schema!.getSchemaFromAdoType(adoType);
-    return schema
+  const codeId = await queryCodeId(address);
+
+  // Try to create a fallback type from ado types query
+  let fallbackType: string | undefined = undefined;
+  const adoType = await client.ado.getType(address).catch(() => undefined);
+  if (adoType) {
+    // Trying to prevent unncessary call for version if type query already failed
+    fallbackType = adoType && await client.ado.getVersion(address).then(version => `${adoType}@${version}`).catch(() => undefined);
   }
+  const schema = await client.schema!.getSchemaFromCodeId(codeId, undefined, fallbackType);
+  return schema;
+
 }
 
 /**
@@ -319,7 +287,6 @@ async function executeHandler(input: string[], flags: Flags) {
     adoSchema.schema.execute
       ? adoSchema.schema.execute
       : (adoSchema.schema as Schema),
-    adoSchema.key
   );
   await executeMessage(address, msg, flags);
 }
@@ -342,14 +309,13 @@ async function queryHandler(input: string[]) {
   const adoSchema = await displaySpinnerAsync(
     "Fetching schema...",
     async () =>
-      await client!.os!.schema!.getSchemaFromCodeId(codeId)
+      await client!.schema!.getSchemaFromCodeId(codeId)
   );
 
   const msg = await promptQueryOrExecuteMessage(
     adoSchema.schema.query
       ? adoSchema.schema.query
       : adoSchema.schema as Schema,
-    adoSchema.key
   );
   const resp = await queryMessage(address, msg);