+++ packages/andrjs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @andromedaprotocol/andromeda.js
 
+## 0.3.8-beta.1
+
+### Patch Changes
+
+- Cosmjs updates, Injective Client Update, ADO Schema updates
+
+## 0.3.8-beta.0
+
+### Patch Changes
+
+- Cosmjs updates, Injective Client Update, ADO Schema updates
+- Updated dependencies
+  - @andromedaprotocol/gql@0.3.8-beta.0
+
 ## 0.3.7
 
 ### Patch Changes