+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @andromedaprotocol/cli
 
+## 0.3.9-beta.1
+
+### Patch Changes
+
+- Cosmjs updates, Injective Client Update, ADO Schema updates
+- Updated dependencies
+  - @andromedaprotocol/andromeda.js@0.3.8-beta.1
+
+## 0.3.9-beta.0
+
+### Patch Changes
+
+- Cosmjs updates, Injective Client Update, ADO Schema updates
+- Updated dependencies
+  - @andromedaprotocol/andromeda.js@0.3.8-beta.0
+
 ## 0.3.7
 
 ### Patch Changes