+++ packages/gql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @andromedaprotocol/gql
 
+## 0.3.8-beta.0
+
+### Patch Changes
+
+- Cosmjs updates, Injective Client Update, ADO Schema updates
+
 ## 0.3.7
 
 ### Patch Changes